data/influx/src/requirements.txt
pytest==6.1.2
pytest-cov==3.0.0
Flask==2.0.2
flake8==4.0.1
mypy==0.931
pylint==2.12.2
isort==5.10.1
safety==1.10.3
gunicorn==20.1.0
python-decouple==3.5
influxdb-client==1.25.0
functions-framework==3.0.0
pytest-mock==3.7.0
